Fix pybroccoli record instantiation.
Declaring more than one record_type could cause it to break because
only the last-declared record_type was used in the instantiation.

The RecTypes global seems like it was meant to hold identifiers for
each record_type that is declared, but instead each record_type just
replaces a single entry in the dictionary keyed with 'record_type'.
This table doesn't seem necessary at all, though, since the record_type
to use for instantiation is directly available to the factory.