In broccoli.py, import from broccoli_intern not _broccoli_intern.
Using swig 3.0.5, the low-level wrapper library (_broccoli_intern.so)
doesn't contain global constants if imported by itself, but the python
wrapper module (broccoli_intern.py) will and seems to also insert them
in to the _broccoli_intern namespace.

Generally, I think it's more typical to import the generated python
module which takes care of loading the wrapper library, so not sure why
it was this way.  And this change should still work on old versions of
swig (tested on 1.3.40).